Refactor Options
