+++ src/main/kotlin/org/olafneumann/regex/generator/model/PatternRecognizerModel.kt
@@ -4,6 +4,7 @@ import dev.andrewbailey.diff.differenceOf
 import org.olafneumann.regex.generator.regex.RegexMatchCombiner
 import org.olafneumann.regex.generator.recognizer.RecognizerMatch
 import org.olafneumann.regex.generator.recognizer.RecognizerRegistry
+import org.olafneumann.regex.generator.regex.Options
 import org.olafneumann.regex.generator.regex.RegularExpression
 import org.olafneumann.regex.generator.util.hasIntersectionWith
 
@@ -11,7 +12,7 @@ data class PatternRecognizerModel(
     val input: String,
     val recognizerMatches: List<RecognizerMatch> = RecognizerRegistry.findMatches(input),
     val selectedRecognizerMatches: Collection<RecognizerMatch> = emptySet(),
-    val options: RegexMatchCombiner.Options,
+    val options: Options,
     val regularExpression: RegularExpression = RegexMatchCombiner
         .combineMatches(inputText = input, selectedMatches = selectedRecognizerMatches, options = options)
 ) {
@@ -75,7 +76,7 @@ data class PatternRecognizerModel(
         )
     }
 
-    fun setOptions(options: RegexMatchCombiner.Options) : PatternRecognizerModel {
+    fun setOptions(options: Options) : PatternRecognizerModel {
         return copy(
             options = options,
             regularExpression = RegexMatchCombiner
