+++ src/main/kotlin/org/olafneumann/regex/generator/output/CRegexCodeGenerator.kt
@@ -1,7 +1,7 @@
-package org.olafneumann.regex.generator.output;
+package org.olafneumann.regex.generator.output
 
 import org.olafneumann.regex.generator.regex.RegexCache
-import org.olafneumann.regex.generator.regex.RegexMatchCombiner
+import org.olafneumann.regex.generator.regex.Options
 
 internal class CRegexCodeGenerator : SimpleReplacingCodeGenerator(
     languageName = "C (regex.h)",
@@ -40,10 +40,10 @@ int useRegex(char* textToCheck) {
     return actualReturnValue;
 }"""
 ) {
-    override fun transformPattern(pattern: String, options: RegexMatchCombiner.Options): String =
+    override fun transformPattern(pattern: String, options: Options): String =
         pattern.replace(RegexCache.get("([\\\\\"])"), "\\\\$1").replace(RegexCache.get("\t"), "\\t")
 
-    override fun generateOptionsCode(options: RegexMatchCombiner.Options) =
+    override fun generateOptionsCode(options: Options) =
         options.combine(
             valueForCaseInsensitive = "REG_ICASE",
             valueForMultiline = "REG_NEWLINE",
@@ -51,7 +51,7 @@ int useRegex(char* textToCheck) {
             prefix = " | "
         )
 
-    override fun getWarnings(pattern: String, options: RegexMatchCombiner.Options): List<String> {
+    override fun getWarnings(pattern: String, options: Options): List<String> {
         if (options.dotMatchesLineBreaks)
             return listOf("DOT_ALL is not supported by regex.h")
