+++ src/main/kotlin/org/olafneumann/regex/generator/output/CSharpCodeGenerator.kt
@@ -1,7 +1,7 @@
 package org.olafneumann.regex.generator.output
 
 import org.olafneumann.regex.generator.regex.RegexCache
-import org.olafneumann.regex.generator.regex.RegexMatchCombiner
+import org.olafneumann.regex.generator.regex.Options
 
 internal class CSharpCodeGenerator : SimpleReplacingCodeGenerator(
     languageName = "C#",
@@ -20,10 +20,10 @@ public class Sample
 ) {
 
 
-    override fun transformPattern(pattern: String, options: RegexMatchCombiner.Options): String =
+    override fun transformPattern(pattern: String, options: Options): String =
         pattern.replace(RegexCache.get("([\\\\\"])"), "\\\\$1").replace(RegexCache.get("\t"), "\\t")
 
-    override fun generateOptionsCode(options: RegexMatchCombiner.Options) =
+    override fun generateOptionsCode(options: Options) =
         options.combine(
             valueForCaseInsensitive = "RegexOptions.IgnoreCase",
             valueForMultiline = "RegexOptions.Multiline",
