+++ src/main/kotlin/org/olafneumann/regex/generator/output/JavaScriptCodeGenerator.kt
@@ -1,6 +1,6 @@
 package org.olafneumann.regex.generator.output;
 
-import org.olafneumann.regex.generator.regex.RegexMatchCombiner
+import org.olafneumann.regex.generator.regex.Options
 
 internal class JavaScriptCodeGenerator : SimpleReplacingCodeGenerator(
     languageName = "JavaScript",
@@ -11,13 +11,13 @@ internal class JavaScriptCodeGenerator : SimpleReplacingCodeGenerator(
 }"""
 ) {
 
-    override fun transformPattern(pattern: String, options: RegexMatchCombiner.Options): String =
+    override fun transformPattern(pattern: String, options: Options): String =
         pattern.replace("/", "\\/").replace("\t", "\\t")
 
-    override fun generateOptionsCode(options: RegexMatchCombiner.Options) =
+    override fun generateOptionsCode(options: Options) =
         options.combine(valueForCaseInsensitive = "i", valueForMultiline = "m", valueForDotAll = "s")
 
-    override fun getWarnings(pattern: String, options: RegexMatchCombiner.Options): List<String> {
+    override fun getWarnings(pattern: String, options: Options): List<String> {
         if (options.dotMatchesLineBreaks)
             return listOf("The option 's' (dot matches line breaks) is not supported in Firefox and IE.")
         return emptyList()
