+++ src/main/kotlin/org/olafneumann/regex/generator/output/RubyCodeGenerator.kt
@@ -1,7 +1,7 @@
 package org.olafneumann.regex.generator.output
 
+import org.olafneumann.regex.generator.regex.Options
 import org.olafneumann.regex.generator.regex.RegexCache
-import org.olafneumann.regex.generator.regex.RegexMatchCombiner
 
 internal class RubyCodeGenerator : SimpleReplacingCodeGenerator(
     languageName = "Ruby",
@@ -12,10 +12,10 @@ internal class RubyCodeGenerator : SimpleReplacingCodeGenerator(
 end"""
 ) {
 
-    override fun transformPattern(pattern: String, options: RegexMatchCombiner.Options): String =
+    override fun transformPattern(pattern: String, options: Options): String =
         pattern.replace(RegexCache.get("([\\\\'])"), "\\\\$1").replace(RegexCache.get("\t"), "\\t")
 
-    override fun generateOptionsCode(options: RegexMatchCombiner.Options) =
+    override fun generateOptionsCode(options: Options) =
         options.combine(
             valueForCaseInsensitive = "Regexp::IGNORECASE",
             valueForDotAll = "Regexp::MULTILINE",
@@ -23,7 +23,7 @@ end"""
             prefix = ", "
         )
 
-    override fun getWarnings(pattern: String, options: RegexMatchCombiner.Options): List<String> {
+    override fun getWarnings(pattern: String, options: Options): List<String> {
         val warnings = mutableListOf<String>()
         if (options.multiline)
             warnings += "The Ruby regex engine does not support the MULTILINE option. " +
