+++ src/main/kotlin/org/olafneumann/regex/generator/output/SwiftCodeGenerator.kt
@@ -1,7 +1,7 @@
 package org.olafneumann.regex.generator.output
 
+import org.olafneumann.regex.generator.regex.Options
 import org.olafneumann.regex.generator.regex.RegexCache
-import org.olafneumann.regex.generator.regex.RegexMatchCombiner
 
 internal class SwiftCodeGenerator : SimpleReplacingCodeGenerator(
     languageName = "Swift",
@@ -14,10 +14,10 @@ internal class SwiftCodeGenerator : SimpleReplacingCodeGenerator(
 }"""
 ) {
 
-    override fun transformPattern(pattern: String, options: RegexMatchCombiner.Options): String =
+    override fun transformPattern(pattern: String, options: Options): String =
         pattern.replace(RegexCache.get("([\\\\\"])"), "\\\\$1").replace(RegexCache.get("\t"), "\\t")
 
-    override fun generateOptionsCode(options: RegexMatchCombiner.Options) =
+    override fun generateOptionsCode(options: Options) =
         options.combine(
             valueForCaseInsensitive = ".caseInsensitive",
             valueForDotAll = ".dotMatchesLineSeparators",
