+++ src/main/kotlin/org/olafneumann/regex/generator/output/UrlGenerator.kt
@@ -1,7 +1,7 @@
 package org.olafneumann.regex.generator.output
 
 import org.olafneumann.regex.generator.js.encodeURIComponent
-import org.olafneumann.regex.generator.regex.RegexMatchCombiner
+import org.olafneumann.regex.generator.regex.Options
 
 internal open class UrlGenerator(
     linkName: String,
@@ -10,10 +10,10 @@ internal open class UrlGenerator(
     private val valueForDotAll: String? = "s",
     private val valueForMultiline: String? = "m",
 ) : SimpleReplacingCodeGenerator(linkName, linkName, urlTemplate) {
-    override fun transformPattern(pattern: String, options: RegexMatchCombiner.Options): String =
+    override fun transformPattern(pattern: String, options: Options): String =
         encodeURIComponent(pattern)
 
-    override fun generateOptionsCode(options: RegexMatchCombiner.Options): String =
+    override fun generateOptionsCode(options: Options): String =
         options.combine(
             valueForCaseInsensitive = valueForCaseInsensitive,
             valueForMultiline = valueForMultiline,
