+++ src/main/kotlin/org/olafneumann/regex/generator/ui/components/LinkHandler.kt
@@ -1,14 +1,14 @@
 package org.olafneumann.regex.generator.ui.components
 
 import org.olafneumann.regex.generator.output.CodeGenerator
-import org.olafneumann.regex.generator.regex.RegexMatchCombiner
+import org.olafneumann.regex.generator.regex.Options
 import org.w3c.dom.HTMLAnchorElement
 
 internal class LinkHandler(
     private val link: HTMLAnchorElement,
     private val codeGenerator: CodeGenerator
 ) {
-    fun setPattern(pattern: String, options: RegexMatchCombiner.Options) {
+    fun setPattern(pattern: String, options: Options) {
         link.href = codeGenerator.generateCode(pattern, options).snippet
     }
 }
