+++ src/main/kotlin/org/olafneumann/regex/generator/ui/parts/PXOptions.kt
@@ -1,7 +1,7 @@
 package org.olafneumann.regex.generator.ui.parts
 
+import org.olafneumann.regex.generator.regex.Options
 import org.olafneumann.regex.generator.ui.model.DisplayModel
-import org.olafneumann.regex.generator.regex.RegexMatchCombiner
 import org.olafneumann.regex.generator.ui.HtmlView
 import org.olafneumann.regex.generator.ui.MVCContract
 import org.olafneumann.regex.generator.ui.utils.HtmlHelper
@@ -24,8 +24,8 @@ class PXOptions(
         checkWholeLine.oninput = { propagateOptionsChange() }
     }
 
-    private var options: RegexMatchCombiner.Options
-        get() = RegexMatchCombiner.Options(
+    private var options: Options
+        get() = Options(
             onlyPatterns = checkOnlyMatches.checked,
             matchWholeLine = checkWholeLine.checked,
             caseInsensitive = checkCaseInsensitive.checked,
